MLO_P5_Sources/requirements.txt
pandas>=2.2.2
pymongo>=4.7.2
python-dotenv>=1.0.1
pytest>=8.3.2
mongomock>=4.1.2
tqdm>=4.66.4
